Migrate reCAPTCHA support to Google reCAPTCHA Enterprise (v2 and v3)
The RegistrationRecaptcha logic is old and targets the first version
of Google reCAPTCHA, which is no longer supported.

This commit updates the logic to align with the new **Google reCAPTCHA
Enterprise**, with support for both v2 and v3 captchas:

- v2: checkbox-based, visible captcha that is displayed on the page
- v3: score-based, invisible captcha that is hidden on the page, and
  based on the user's behavior and identity on other sites.

The old configuration (site key, secret key) is replaced with
project ID, site key and API key.

Two new optional configurations are:

- a "v3" toggle: there is no easy way to deduce from a site key which
  version is enabled. The user hence has to provide this information
  to the widget.
- a "min threshold": this is the minimal score a captcha must return
  for the registration to proceed. It should be between 0.0 and 1.0,
  and is set to the sensible default of 0.7 - "medium high".

The `register.ftl` template is updated as well. Note that in case of
invisible captcha, a call to the captcha API must be made to retrieve
the captcha token before submitting the form.
This and the new theme attributes will require custom theme maintainer
to udpate their code.

The documentation is also updated to reflect the changes.

*About the new reCAPTCHA*

The new reCAPTCHA version works in multiple stages:

1. on the frontend, a JSON script is loaded on the page, and:
   * with v2, an input (populated automatically) is added to the form,
     with the checkbox,
   * with v3, nothing is shown, so it is up to us to request a captcha
     token before submitting the form. This can be done with a button
     and a callback.
2. on the backend, an assessment must be created and sent to google,
   along with the captcha token (this is where the API key is necessary),
3. the google assessment response must then be interpreted, to decide if
   the reCAPTCHA is valid. This entails checking the action name, the "valid"
   attribute and the score.

Closes: #16138